requirements.txt
pyserial==3.5
dataclasses==0.8
